requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
python-dotenv==1.0.0
boto3>=1.34.72
botocore>=1.34.72
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
bcrypt==4.1.2
sqlalchemy==2.0.23
sqlite-utils==3.35.2
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.26.2
pandas==2.1.3
langchain==0.1.0
langchain-community==0.0.10
langchain-aws==0.1.0
Pillow==10.1.0
email-validator==2.1.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0

